Securing apis

+++ server/lib/components/routes/accounts.js
@@ -11,6 +11,8 @@ export default function(options = {}) {
 
     app.get('/api/accounts', async (req, res, next) => {
       try {
+        if (!req.user.hasPermission('accounts-read')) return next(Boom.forbidden());
+
         const limit = req.query.limit ? parseInt(req.query.limit, 10) : undefined;
         const offset = req.query.offset ? parseInt(req.query.offset, 10) : undefined;
         const result = await store.findAccounts({}, limit, offset);
@@ -22,6 +24,8 @@ export default function(options = {}) {
 
     app.get('/api/accounts/:id', async (req, res, next) => {
       try {
+        if (!req.user.hasPermissionOnAccount(req.params.id, 'accounts-read')) return next(Boom.forbidden());
+
         const account = await store.getAccount(req.params.id);
         return account ? res.json(account) : next();
       } catch (err) {
@@ -43,6 +47,8 @@ export default function(options = {}) {
 
     app.delete('/api/accounts/:id', async (req, res, next) => {
       try {
+        if (!req.user.hasPermissionOnAccount(req.params.id, 'accounts-write')) return next(Boom.forbidden());
+
         const meta = { date: new Date(), account: { id: req.user.id, }, };
         await store.deleteAccount(req.params.id, meta);
         res.status(204).send();
@@ -58,6 +64,8 @@ export default function(options = {}) {
         if (!req.body.provider) return next(Boom.badRequest('provider is required'));
         if (!req.body.type) return next(Boom.badRequest('type is required'));
 
+        if (!req.user.hasPermissionOnAccount(req.body.account, 'accounts-write')) return next(Boom.forbidden());
+
         const data = { name: req.body.name, provider: req.body.provider, type: req.body.type, };
         const meta = { date: new Date(), account: { id: req.user.id, }, };
         const identity = await store.saveIdentity(req.body.account, data, meta);
@@ -67,22 +75,14 @@ export default function(options = {}) {
       }
     });
 
-    app.delete('/api/identities/:id', async (req, res, next) => {
-      try {
-        await store.deleteIdentity(req.params.id, { date: new Date(), account: req.user.id, });
-        res.status(204).send();
-      } catch (err) {
-        next(err);
-      }
-    });
-
     app.post('/api/roles/registry', bodyParser.json(), async (req, res, next) => {
-
-      if (!req.body.account) return next(Boom.badRequest('account is required'));
-      if (!req.body.role) return next(Boom.badRequest('role is required'));
-      if (!req.body.registry) return next(Boom.badRequest('registry is required'));
-
       try {
+        if (!req.body.account) return next(Boom.badRequest('account is required'));
+        if (!req.body.role) return next(Boom.badRequest('role is required'));
+        if (!req.body.registry) return next(Boom.badRequest('registry is required'));
+
+        if (!req.user.hasPermissionOnRegistry(req.body.registry, 'registries-grant')) return next(Boom.forbidden());
+
         const meta = { date: new Date(), account: { id: req.user.id, }, };
         const account = await store.grantRoleOnRegistry(req.body.account, req.body.role, req.body.registry, meta);
         res.json(account);
@@ -92,12 +92,13 @@ export default function(options = {}) {
     });
 
     app.post('/api/roles/namespace', bodyParser.json(), async (req, res, next) => {
-
-      if (!req.body.account) return next(Boom.badRequest('account is required'));
-      if (!req.body.role) return next(Boom.badRequest('role is required'));
-      if (!req.body.namespace) return next(Boom.badRequest('namespace is required'));
-
       try {
+        if (!req.body.account) return next(Boom.badRequest('account is required'));
+        if (!req.body.role) return next(Boom.badRequest('role is required'));
+        if (!req.body.namespace) return next(Boom.badRequest('namespace is required'));
+
+        if (!req.user.hasPermissionOnNamespace(req.body.namespace, 'namespaces-grant')) return next(Boom.forbidden());
+
         const meta = { date: new Date(), account: { id: req.user.id, }, };
         const account = await store.grantRoleOnNamespace(req.body.account, req.body.role, req.body.namespace, meta);
         res.json(account);
@@ -105,25 +106,6 @@ export default function(options = {}) {
         next(err);
       }
     });
-
-    app.delete('/api/roles/registry/:id', async (req, res, next) => {
-      try {
-        await store.revokeRoleOnRegistry(req.params.id, { date: new Date(), account: req.user.id, });
-        res.status(204).send();
-      } catch (err) {
-        next(err);
-      }
-    });
-
-    app.delete('/api/roles/namespace/:id', async (req, res, next) => {
-      try {
-        await store.revokeRoleOnNamespace(req.params.id, { date: new Date(), account: req.user.id, });
-        res.status(204).send();
-      } catch (err) {
-        next(err);
-      }
-    });
-
 
     cb();
   }
